feat(spell-check): introduce new bash script for spell check (#8)
* feat(spell-check): introduce new bash script for spell check
✨This commit introduces a new bash file titled 'spell-check.bash'. The bash script executes a spell check over .tex files, specifically checking for errors in the German language. Highlights of the script include color-coded error messages and sorting of unique spelling errors.

Additionally, this commit adds a new 'company.tex' file under the chapters/BPP directory. The file contains details about the firm SDA SE, its timeline, corporate structure, product spectrum, economic indicators, organizational structure and the personal tasks of Mr. Siemer.

🚧 While introducing the bash script, we found that it doesn't handle file names with spaces or special characters well. Future updates will target this issue.

🔧Further debugging and optimization of the bash script are planned. This includes refining the color-coding and error messages for easier comprehension, and expanding the scope to other file types and languages.

📚The 'company.tex' file

* feat(chapters): update 'BPP/company.tex' with simplified content
💡 Simplify and clarify the content in 'BPP/company.tex'
The changes include a more concise description of:
- The company's history
- Its current status and objectives
- The roles and responsibilities of key personnel

Additionally, the text has been restructured for better flow and readability.
Remember, it's not the code that's complex. It's the business that it supports. 🎉

In this update, I've tried to ensure that the content mirrors the
philosophy of simplicity and efficiency that we cherish in our
development practices. 🚀 Just like Bill says, "The advance of technology is based on making it fit in so
that you don't really even notice it, so it's part of everyday life." 🕹️

Cheers to making complexity look simple and daunting tasks more manageable! 🍻

Signed-off-by: Karsten Siemer <karsten.siemer@sda.se>

* feat(thesis): revise and refine content
✍️ Revamp the text and structure in company.tex and thesis.tex files.
👨‍💻 Some humour element - Feels like I'm back to my Microsoft days, adding
some break points in the text for readability instead of coding!

Changes involve:
- Sprucing up the document by adding some "/medskip" and "/noindent" for
better structure and readability.
- Tweaking the narrative tone from first person to third person.
- Renaming the part title from "Thesis" to "Abschlussbericht" in
thesis.tex.
+1 for trying to keep the readers awake with these changes! 🤓

May the force of comprehensive writing be with you!

---------

Signed-off-by: Karsten Siemer <karsten.siemer@sda.se>